put uuid with CRLF in windows builds
